+++ back/fastats/src/main/java/org/sixbacks/fastats/statistics/repository/StatSurveyRepository.java
@@ -3,6 +3,6 @@ package org.sixbacks.fastats.statistics.repository;
 import org.sixbacks.fastats.statistics.entity.StatSurvey;
 import org.springframework.data.repository.ListCrudRepository;
 
-public interface StatSurveyRepository extends ListCrudRepository<StatSurvey, Integer> {
+public interface StatSurveyRepository extends ListCrudRepository<StatSurvey, Long> {
 
 }
